refactored visualization
